code/llmxive-automation/helpers/requirements.txt
# Core dependencies
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0
huggingface-hub>=0.19.0
PyGithub>=2.1.0
click>=8.1.0

# Model optimization
bitsandbytes>=0.41.0  # For 8-bit quantization
einops>=0.7.0

# Optional model support
# llama-cpp-python>=0.2.0  # For GGUF models (uncomment if needed)
# auto-gptq>=0.5.0  # For GPTQ models (uncomment if needed)

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0